# maquinaria_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Maquinaria",
        "description": "Documentación de la API de Mantenimientos y Maquinarias",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/representantes": {
            "get": {
                "tags": [
                    "Representantes"
                ],
                "summary": "Listar representantes",
                "operationId": "6008b663f14f39a1522320f3f85bab7a",
                "responses": {
                    "200": {
                        "description": "Lista de representantes con su empresa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Representantes"
                ],
                "summary": "Crear un nuevo representante",
                "operationId": "0864af7e6f3dfda79237aa41a3e2bce5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "documento",
                                    "telefono",
                                    "email",
                                    "empresa_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001234567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "empresa_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Representante creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/representantes/{id}": {
            "get": {
                "tags": [
                    "Representantes"
                ],
                "summary": "Obtener un representante",
                "operationId": "e41e6b27b54ada68ee6c9b87f6df3aea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del representante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Representante encontrado"
                    },
                    "404": {
                        "description": "Representante no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Representantes"
                ],
                "summary": "Actualizar un representante",
                "operationId": "9416ee2beb70668005c7776a3fee5696",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del representante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Ramírez"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "example": "98765432"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3209876543"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "carlos@example.com"
                                    },
                                    "empresa_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Representante actualizado exitosamente"
                    },
                    "404": {
                        "description": "Representante no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Representantes"
                ],
                "summary": "Eliminar un representante",
                "operationId": "483d2e61f4384d2135c7c828095e9a98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del representante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Representante eliminado exitosamente"
                    },
                    "404": {
                        "description": "Representante no encontrado"
                    }
                }
            }
        },
...
    },
...
}